refactor: Domain 과 repository 의 패키지를 분리합니다
