+++ src/main/java/nextstep/subway/applicaion/SubwayLineService.java
@@ -3,14 +3,14 @@ package nextstep.subway.applicaion;
 import nextstep.subway.applicaion.dto.subwayLine.CreateSubwayLineRequest;
 import nextstep.subway.applicaion.dto.subwayLine.SubwayLineResponse;
 import nextstep.subway.applicaion.dto.subwayLine.UpdateSubwayLineRequest;
-import nextstep.subway.domain.stationToSubwayBridge.StationToSubwayLine;
-import nextstep.subway.domain.stationToSubwayBridge.StationToSubwayLineRepository;
-import nextstep.subway.domain.station.Station;
-import nextstep.subway.domain.station.StationRepository;
-import nextstep.subway.domain.subwayLine.SubwayLine;
-import nextstep.subway.domain.subwayLine.SubwayLineRepository;
-import nextstep.subway.domain.subwayLineColor.SubwayLineColor;
-import nextstep.subway.domain.subwayLineColor.SubwayLineColorRepository;
+import nextstep.subway.domain.StationToSubwayLine;
+import nextstep.subway.repository.StationToSubwayLineRepository;
+import nextstep.subway.domain.Station;
+import nextstep.subway.repository.StationRepository;
+import nextstep.subway.domain.SubwayLine;
+import nextstep.subway.repository.SubwayLineRepository;
+import nextstep.subway.domain.SubwayLineColor;
+import nextstep.subway.repository.SubwayLineColorRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
