+++ src/main/java/nextstep/subway/repository/StationRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain.station;
+package nextstep.subway.repository;
 
+import nextstep.subway.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
