+++ src/main/java/nextstep/subway/repository/SubwayLineColorRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain.subwayLineColor;
+package nextstep.subway.repository;
 
+import nextstep.subway.domain.SubwayLineColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
